--- a/broker_redis.go
+++ b/broker_redis.go
@@ -13,9 +13,10 @@ import (
 	"time"
 
 	"github.com/centrifugal/centrifuge/internal/timers"
+	"github.com/centrifugal/centrifuge/internal/util"
 
 	"github.com/centrifugal/protocol"
-	"github.com/gomodule/redigo/redis"
+	"github.com/go-redis/redis/v8"
 )
 
 const (
@@ -127,10 +128,10 @@ func NewRedisBroker(n *Node, config RedisBrokerConfig) (*RedisBroker, error) {
 		shards:                 config.Shards,
 		config:                 config,
 		sharding:               len(config.Shards) > 1,
-		historyListScript:      redis.NewScript(2, historyListSource),
-		historyStreamScript:    redis.NewScript(2, historyStreamSource),
-		addHistoryListScript:   redis.NewScript(2, addHistorySource),
-		addHistoryStreamScript: redis.NewScript(2, addHistoryStreamSource),
+		historyListScript:      redis.NewScript(historyListSource),
+		historyStreamScript:    redis.NewScript(historyStreamSource),
+		addHistoryListScript:   redis.NewScript(addHistorySource),
+		addHistoryStreamScript: redis.NewScript(addHistoryStreamSource),
 		closeCh:                make(chan struct{}),
 	}
 
@@ -347,7 +348,7 @@ func (b *RedisBroker) checkCapabilities(shard *RedisShard) error {
 		return nil
 	}
 	// Check whether Redis Streams supported.
-	dr := shard.newDataRequest("XRANGE", nil, "", []interface{}{"_", "0-0", "0-0"})
+	dr := shard.newDataRequest(nil, nil, []interface{}{"XRANGE", "_", "0-0", "0-0"})
 	resp := shard.getDataResponse(dr, b.closeCh)
 	if resp.err != nil {
 		if strings.Contains(resp.err.Error(), "ERR unknown command") {
@@ -437,7 +438,7 @@ func (b *RedisBroker) publish(s *RedisShard, ch string, data []byte, opts Publis
 		size = opts.HistorySize - 1
 		script = b.addHistoryListScript
 	}
-	dr := s.newDataRequest("", script, streamKey, []interface{}{streamKey, historyMetaKey, byteMessage, size, int(opts.HistoryTTL.Seconds()), publishChannel, historyMetaTTLSeconds, time.Now().Unix()})
+	dr := s.newDataRequest(script, []string{string(streamKey), string(historyMetaKey)}, []interface{}{byteMessage, size, int(opts.HistoryTTL.Seconds()), string(publishChannel), historyMetaTTLSeconds, time.Now().Unix()})
 	resp := s.getDataResponse(dr, b.closeCh)
 	if resp.err != nil {
 		return StreamPosition{}, resp.err
@@ -446,11 +447,13 @@ func (b *RedisBroker) publish(s *RedisShard, ch string, data []byte, opts Publis
 	if !ok || len(replies) != 2 {
 		return StreamPosition{}, errors.New("wrong Redis reply")
 	}
-	offset, err := redis.Uint64(replies[0], nil)
+	r := redis.NewCmdResult(replies[0], nil)
+	offset, err := r.Uint64()
 	if err != nil {
 		return StreamPosition{}, errors.New("wrong Redis reply offset")
 	}
-	epoch, err := redis.String(replies[1], nil)
+	r.SetVal(replies[1])
+	epoch, err := r.Text()
 	if err != nil {
 		return StreamPosition{}, errors.New("wrong Redis reply epoch")
 	}
@@ -631,8 +634,9 @@ func (b *RedisBroker) removeHistory(s *RedisShard, ch string) error {
 	} else {
 		key = b.historyListKey(s, ch)
 	}
-	dr := s.newDataRequest("DEL", nil, key, []interface{}{key})
+	dr := s.newDataRequest(nil, nil, []interface{}{"DEL", string(key)})
 	resp := s.getDataResponse(dr, b.closeCh)
+
 	return resp.err
 }
 
@@ -679,21 +683,8 @@ func (b *RedisBroker) runPubSub(s *RedisShard, eventHandler BrokerEventHandler) 
 		b.node.Log(NewLogEntry(LogLevelDebug, "stopping Redis PUB/SUB", map[string]interface{}{"shard": s.string()}))
 	}()
 
-	poolConn := s.pool.Get()
-	if poolConn.Err() != nil {
-		// At this moment test on borrow could already return an error,
-		// we can't work with broken connection.
-		_ = poolConn.Close()
-		return
-	}
-
-	var connMu sync.Mutex
-	conn := redis.PubSubConn{Conn: poolConn}
-	defer func() {
-		connMu.Lock()
-		_ = conn.Close()
-		connMu.Unlock()
-	}()
+	pubsub := s.client.Subscribe(context.Background())
+	defer func() { _ = pubsub.Close() }()
 
 	done := make(chan struct{})
 	var doneOnce sync.Once
@@ -716,10 +707,7 @@ func (b *RedisBroker) runPubSub(s *RedisShard, eventHandler BrokerEventHandler) 
 	// Run subscriber goroutine.
 	go func() {
 		defer func() {
-			connMu.Lock()
-			// Error here automatically closes a connection in Redigo.
-			_ = conn.Unsubscribe()
-			connMu.Unlock()
+			_ = pubsub.Unsubscribe(context.Background())
 		}()
 		b.node.Log(NewLogEntry(LogLevelDebug, "starting RedisBroker Subscriber", map[string]interface{}{"shard": s.string()}))
 		defer func() {
@@ -736,9 +724,9 @@ func (b *RedisBroker) runPubSub(s *RedisShard, eventHandler BrokerEventHandler) 
 				isSubscribe := r.subscribe
 				channelBatch := []subRequest{r}
 
-				chIDs := make([]interface{}, 0, len(r.channels))
+				chIDs := make([]string, 0, len(r.channels))
 				for _, ch := range r.channels {
-					chIDs = append(chIDs, ch)
+					chIDs = append(chIDs, string(ch))
 				}
 
 				var otherR *subRequest
@@ -756,7 +744,7 @@ func (b *RedisBroker) runPubSub(s *RedisShard, eventHandler BrokerEventHandler) 
 						}
 						channelBatch = append(channelBatch, r)
 						for _, ch := range r.channels {
-							chIDs = append(chIDs, ch)
+							chIDs = append(chIDs, string(ch))
 						}
 					default:
 						break loop
@@ -765,9 +753,9 @@ func (b *RedisBroker) runPubSub(s *RedisShard, eventHandler BrokerEventHandler) 
 
 				var opErr error
 				if isSubscribe {
-					opErr = conn.Subscribe(chIDs...)
+					opErr = pubsub.Subscribe(context.Background(), chIDs...)
 				} else {
-					opErr = conn.Unsubscribe(chIDs...)
+					opErr = pubsub.Unsubscribe(context.Background(), chIDs...)
 				}
 
 				if opErr != nil {
@@ -784,15 +772,15 @@ func (b *RedisBroker) runPubSub(s *RedisShard, eventHandler BrokerEventHandler) 
 					r.done(nil)
 				}
 				if otherR != nil {
-					chIDs := make([]interface{}, 0, len(otherR.channels))
+					chIDs := make([]string, 0, len(otherR.channels))
 					for _, ch := range otherR.channels {
-						chIDs = append(chIDs, ch)
+						chIDs = append(chIDs, string(ch))
 					}
 					var opErr error
 					if otherR.subscribe {
-						opErr = conn.Subscribe(chIDs...)
+						opErr = pubsub.Subscribe(context.Background(), chIDs...)
 					} else {
-						opErr = conn.Unsubscribe(chIDs...)
+						opErr = pubsub.Unsubscribe(context.Background(), chIDs...)
 					}
 					if opErr != nil {
 						otherR.done(opErr)
@@ -806,11 +794,11 @@ func (b *RedisBroker) runPubSub(s *RedisShard, eventHandler BrokerEventHandler) 
 	}()
 
 	// Run workers to spread received message processing work over worker goroutines.
-	workers := make(map[int]chan redis.Message)
+	workers := make(map[int]chan *redis.Message)
 	for i := 0; i < numWorkers; i++ {
-		workerCh := make(chan redis.Message, redisPubSubWorkerChannelSize)
+		workerCh := make(chan *redis.Message, redisPubSubWorkerChannelSize)
 		workers[i] = workerCh
-		go func(ch chan redis.Message) {
+		go func(ch chan *redis.Message) {
 			for {
 				select {
 				case <-done:
@@ -820,7 +808,7 @@ func (b *RedisBroker) runPubSub(s *RedisShard, eventHandler BrokerEventHandler) 
 					case b.pingChannel:
 						// Do nothing - this message just maintains connection open.
 					default:
-						err := b.handleRedisClientMessage(eventHandler, channelID(n.Channel), n.Data)
+						err := b.handleRedisClientMessage(eventHandler, channelID(n.Channel), util.StringToBytes(n.Payload))
 						if err != nil {
 							b.node.Log(NewLogEntry(LogLevelError, "error handling client message", map[string]interface{}{"error": err.Error()}))
 							continue
@@ -867,22 +855,24 @@ func (b *RedisBroker) runPubSub(s *RedisShard, eventHandler BrokerEventHandler) 
 	}()
 
 	for {
-		switch n := conn.ReceiveWithTimeout(10 * time.Second).(type) {
-		case redis.Message:
+		m, err := pubsub.ReceiveTimeout(context.Background(), 10*time.Second)
+		if err != nil {
+			b.node.Log(NewLogEntry(LogLevelError, "Redis PUB/SUB error", map[string]interface{}{"error": err.Error()}))
+			s.reloadPipeline()
+			return
+		}
+		switch m := m.(type) {
+		case *redis.Message:
 			// Add message to worker channel preserving message order - i.b. messages
 			// from the same channel will be processed in the same worker.
 			select {
-			case workers[index(n.Channel, numWorkers)] <- n:
+			case workers[index(m.Channel, numWorkers)] <- m:
 			case <-done:
 				return
 			}
-		case error:
-			b.node.Log(NewLogEntry(LogLevelError, "Redis PUB/SUB error", map[string]interface{}{"error": n.Error()}))
-			s.reloadPipeline()
-			return
-		case redis.Subscription:
+		case *redis.Subscription:
 			// Zero count or unsubscribe from pingChannel signal exit from PUB/SUB routine.
-			if n.Count == 0 || (n.Kind == "unsubscribe" && n.Channel == b.pingChannel) {
+			if m.Count == 0 || (m.Kind == "unsubscribe" && m.Channel == b.pingChannel){
 				return
 			}
 		default:
@@ -896,22 +886,6 @@ func (b *RedisBroker) runControlPubSub(s *RedisShard, eventHandler BrokerEventHa
 	b.node.Log(NewLogEntry(LogLevelDebug, fmt.Sprintf("running Redis control PUB/SUB, num workers: %d", numWorkers), map[string]interface{}{"shard": s.string()}))
 	defer func() {
 		b.node.Log(NewLogEntry(LogLevelDebug, "stopping Redis control PUB/SUB", map[string]interface{}{"shard": s.string()}))
-	}()
-
-	poolConn := s.pool.Get()
-	if poolConn.Err() != nil {
-		// At this moment test on borrow could already return an error,
-		// we can't work with broken connection.
-		_ = poolConn.Close()
-		return
-	}
-
-	var connMu sync.Mutex
-	conn := redis.PubSubConn{Conn: poolConn}
-	defer func() {
-		connMu.Lock()
-		_ = conn.Close()
-		connMu.Unlock()
 	}()
 
 	done := make(chan struct{})
@@ -928,7 +902,7 @@ func (b *RedisBroker) runControlPubSub(s *RedisShard, eventHandler BrokerEventHa
 	pingChannel := b.pingChannel
 
 	// Run workers to spread message processing work over worker goroutines.
-	workCh := make(chan redis.Message)
+	workCh := make(chan *redis.Message)
 	for i := 0; i < numWorkers; i++ {
 		go func() {
 			for {
@@ -940,7 +914,7 @@ func (b *RedisBroker) runControlPubSub(s *RedisShard, eventHandler BrokerEventHa
 					case pingChannel:
 						// Do nothing - this message just maintains connection open.
 					default:
-						err := eventHandler.HandleControl(n.Data)
+						err := eventHandler.HandleControl(util.StringToBytes(n.Payload))
 						if err != nil {
 							b.node.Log(NewLogEntry(LogLevelError, "error handling control message", map[string]interface{}{"error": err.Error()}))
 							continue
@@ -951,22 +925,17 @@ func (b *RedisBroker) runControlPubSub(s *RedisShard, eventHandler BrokerEventHa
 		}()
 	}
 
-	connMu.Lock()
-	err := conn.Subscribe(controlChannel, nodeChannel, pingChannel)
+	pubsub := s.client.Subscribe(context.Background())
+	err := pubsub.Subscribe(context.Background(), controlChannel, nodeChannel, pingChannel)
 	if err != nil {
-		connMu.Unlock()
 		b.node.Log(NewLogEntry(LogLevelError, "control channel subscribe error", map[string]interface{}{"error": err.Error()}))
 		closeDoneOnce()
 		return
 	}
-	connMu.Unlock()
 
 	go func() {
 		defer func() {
-			connMu.Lock()
-			// Error here automatically closes a connection in Redigo.
-			_ = conn.Unsubscribe()
-			connMu.Unlock()
+			_ = pubsub.Unsubscribe(context.Background())
 		}()
 		select {
 		case <-b.closeCh:
@@ -977,23 +946,27 @@ func (b *RedisBroker) runControlPubSub(s *RedisShard, eventHandler BrokerEventHa
 	}()
 
 	for {
-		switch n := conn.ReceiveWithTimeout(10 * time.Second).(type) {
-		case redis.Message:
+		m, err := pubsub.ReceiveTimeout(context.Background(), 10*time.Second)
+		if err != nil {
+			b.node.Log(NewLogEntry(LogLevelError, "Redis control PUB/SUB error", map[string]interface{}{"error": err}))
+			_ = pubsub.Close()
+			return
+		}
+		switch m := m.(type) {
+		case *redis.Message:
 			// Add message to worker channel preserving message order - i.b. messages
 			// from the same channel will be processed in the same worker.
 			select {
-			case workCh <- n:
+			case workCh <- m:
 			case <-done:
 				return
 			}
-		case redis.Subscription:
+		case *redis.Subscription:
 			// Zero count or unsubscribe from pingChannel signal exit from PUB/SUB routine.
-			if n.Count == 0 || (n.Kind == "unsubscribe" && n.Channel == b.pingChannel) {
+			if m.Count == 0 || (m.Kind == "unsubscribe" && m.Channel == b.pingChannel){
 				return
 			}
-		case error:
-			b.node.Log(NewLogEntry(LogLevelError, "Redis control PUB/SUB error", map[string]interface{}{"error": n.Error()}))
-			return
+		case *redis.Pong:
 		}
 	}
 }
@@ -1048,6 +1021,7 @@ func (b *RedisBroker) handleRedisClientMessage(eventHandler BrokerEventHandler, 
 func (b *RedisBroker) runPubSubPing(s *RedisShard) {
 	pingTicker := time.NewTicker(time.Second)
 	defer pingTicker.Stop()
+	client := s.client
 	for {
 		select {
 		case <-b.closeCh:
@@ -1055,20 +1029,18 @@ func (b *RedisBroker) runPubSubPing(s *RedisShard) {
 		case <-pingTicker.C:
 			// Publish periodically to maintain PUB/SUB connection alive and allow
 			// PUB/SUB connection to close early if no data received for a period of time.
-			conn := s.pool.Get()
-			err := conn.Send("PUBLISH", b.pingChannel, nil)
-			if err != nil {
-				b.node.Log(NewLogEntry(LogLevelError, "error publish ping to Redis channel", map[string]interface{}{"error": err.Error()}))
-				_ = conn.Close()
+			cmd := client.Publish(context.Background(), b.pingChannel, nil)
+			if cmd.Err() != nil {
+				b.node.Log(NewLogEntry(LogLevelError, "error publish ping to Redis channel", map[string]interface{}{"error": cmd.Err().Error()}))
 				return
 			}
-			_ = conn.Close()
 		}
 	}
 }
 
 func (b *RedisBroker) runPublishPipeline(s *RedisShard) {
 	var prs []pubRequest
+	client := s.client
 
 	for {
 		select {
@@ -1085,30 +1057,24 @@ func (b *RedisBroker) runPublishPipeline(s *RedisShard) {
 					break loop
 				}
 			}
-			conn := s.pool.Get()
-			for i := range prs {
-				_ = conn.Send("PUBLISH", prs[i].channel, prs[i].message)
+		}
+		cmds, err := client.Pipelined(context.Background(), func(pipe redis.Pipeliner) error {
+			for _, pr := range prs {
+				_ = pipe.Publish(context.Background(), string(pr.channel), pr.message)
 			}
-			err := conn.Flush()
-			if err != nil {
-				for i := range prs {
-					prs[i].done(err)
-				}
-				b.node.Log(NewLogEntry(LogLevelError, "error flushing publish pipeline", map[string]interface{}{"error": err.Error()}))
-				_ = conn.Close()
-				return
-			}
+			return nil
+		})
+		if err != nil {
 			for i := range prs {
-				_, err := conn.Receive()
 				prs[i].done(err)
 			}
-			if conn.Err() != nil {
-				_ = conn.Close()
-				return
-			}
-			_ = conn.Close()
-			prs = nil
+			b.node.Log(NewLogEntry(LogLevelError, "error flushing publish pipeline", map[string]interface{}{"error": err.Error()}))
+			return
 		}
+		for i := range prs {
+			prs[i].done(cmds[i].Err())
+		}
+		prs = nil
 	}
 }
 
@@ -1132,15 +1098,17 @@ func (b *RedisBroker) sendSubscribe(s *RedisShard, r subRequest) error {
 func extractHistoryResponse(reply interface{}, useStreams bool, includePubs bool) (StreamPosition, []*Publication, error) {
 	results := reply.([]interface{})
 
-	offset, err := redis.Uint64(results[0], nil)
+	r := redis.NewCmdResult(results[0], nil)
+	offset, err := r.Uint64()
 	if err != nil {
-		if err != redis.ErrNil {
+		if r.Val() != nil {
 			return StreamPosition{}, nil, err
 		}
 		offset = 0
 	}
 
-	epoch, err := redis.String(results[1], nil)
+	r.SetVal(results[1])
+	epoch, err := r.Text()
 	if err != nil {
 		return StreamPosition{}, nil, err
 	}
@@ -1189,7 +1157,7 @@ func (b *RedisBroker) historyStream(s *RedisShard, ch string, filter HistoryFilt
 
 	historyMetaTTLSeconds := int(b.config.HistoryMetaTTL.Seconds())
 
-	dr := s.newDataRequest("", b.historyStreamScript, historyKey, []interface{}{historyKey, historyMetaKey, includePubs, offset, limit, filter.Reverse, historyMetaTTLSeconds, time.Now().Unix()})
+	dr := s.newDataRequest(b.historyStreamScript, []string{string(historyKey), string(historyMetaKey)}, []interface{}{includePubs, offset, limit, filter.Reverse, historyMetaTTLSeconds, time.Now().Unix()})
 	resp := s.getDataResponse(dr, b.closeCh)
 	if resp.err != nil {
 		return nil, StreamPosition{}, resp.err
@@ -1216,7 +1184,7 @@ func (b *RedisBroker) historyList(s *RedisShard, ch string, filter HistoryFilter
 
 	historyMetaTTLSeconds := int(b.config.HistoryMetaTTL.Seconds())
 
-	dr := s.newDataRequest("", b.historyListScript, historyKey, []interface{}{historyKey, historyMetaKey, includePubs, rightBound, historyMetaTTLSeconds, time.Now().Unix()})
+	dr := s.newDataRequest(b.historyListScript, []string{string(historyKey), string(historyMetaKey)}, []interface{}{includePubs, rightBound, historyMetaTTLSeconds, time.Now().Unix()})
 	resp := s.getDataResponse(dr, b.closeCh)
 	if resp.err != nil {
 		return nil, StreamPosition{}, resp.err
@@ -1317,7 +1285,7 @@ func extractPushData(data []byte) ([]byte, pushType, StreamPosition, bool) {
 		return rest, leavePushType, StreamPosition{}, true
 	}
 
-	stringContent := string(content)
+	stringContent := util.BytesToString(content)
 
 	if contentType == 'p' {
 		// new format p1:offset:epoch
@@ -1339,14 +1307,15 @@ func extractPushData(data []byte) ([]byte, pushType, StreamPosition, bool) {
 }
 
 func sliceOfPubsStream(result interface{}, err error) ([]*Publication, error) {
-	values, err := redis.Values(result, err)
+	values, err := redis.NewCmdResult(result, err).Slice()
 	if err != nil {
 		return nil, err
 	}
 	pubs := make([]*Publication, 0, len(values))
 
 	for i := 0; i < len(values); i++ {
-		streamElementValues, err := redis.Values(values[i], nil)
+		r := redis.NewCmdResult(values[i], nil)
+		streamElementValues, err := r.Slice()
 		if err != nil {
 			return nil, err
 		}
@@ -1355,7 +1324,8 @@ func sliceOfPubsStream(result interface{}, err error) ([]*Publication, error) {
 			return nil, errors.New("malformed reply: number of streamElementValues is not 2")
 		}
 
-		offsetStr, err := redis.String(streamElementValues[0], nil)
+		r.SetVal(streamElementValues[0])
+		offsetStr, err := r.Text()
 		if err != nil {
 			return nil, err
 		}
@@ -1368,8 +1338,8 @@ func sliceOfPubsStream(result interface{}, err error) ([]*Publication, error) {
 			return nil, err
 		}
 
-		val := streamElementValues[1]
-		payloadElementValues, err := redis.Values(val, nil)
+		r.SetVal(streamElementValues[1])
+		payloadElementValues, err := r.Slice()
 		if err != nil {
 			return nil, err
 		}
@@ -1378,13 +1348,13 @@ func sliceOfPubsStream(result interface{}, err error) ([]*Publication, error) {
 			return nil, errors.New("malformed reply: number of payloadElementValues less than 2")
 		}
 
-		pushData, ok := payloadElementValues[1].([]byte)
+		pushData, ok := payloadElementValues[1].(string)
 		if !ok {
-			return nil, errors.New("error getting []byte push data")
+			return nil, errors.New("error getting string push data")
 		}
 
 		var pub protocol.Publication
-		err = pub.UnmarshalVT(pushData)
+		err = pub.UnmarshalVT(util.StringToBytes(pushData))
 		if err != nil {
 			return nil, fmt.Errorf("can not unmarshal value to Publication: %v", err)
 		}
@@ -1395,19 +1365,19 @@ func sliceOfPubsStream(result interface{}, err error) ([]*Publication, error) {
 }
 
 func sliceOfPubsList(result interface{}, err error) ([]*Publication, error) {
-	values, err := redis.Values(result, err)
+	values, err := redis.NewCmdResult(result, err).Slice()
 	if err != nil {
 		return nil, err
 	}
 	pubs := make([]*Publication, 0, len(values))
 
 	for i := len(values) - 1; i >= 0; i-- {
-		value, okValue := values[i].([]byte)
+		value, okValue := values[i].(string)
 		if !okValue {
-			return nil, errors.New("error getting Message value")
+			return nil, errors.New("error getting string value")
 		}
 
-		pushData, _, sp, ok := extractPushData(value)
+		pushData, _, sp, ok := extractPushData(util.StringToBytes(value))
 		if !ok {
 			return nil, fmt.Errorf("malformed publication value: %s", value)
 		}