--- a/internal/util/safe.go
+++ b/internal/util/safe.go
@@ -1,0 +1,11 @@
+//go:build appengine
+
+package util
+
+func BytesToString(b []byte) string {
+	return string(b)
+}
+
+func StringToBytes(s string) []byte {
+	return []byte(s)
+}