@FZambia 2022-07-31
Oh, wow, that's the hard work! And seems great overall 👍 Removing dependencies and the code that manages various connection modes is awesome.

Probably let's collaborate next time before introducing such a big changes? Just to make sure we are on the same page about things.

In general, the issue was about researching things, we need to prove the migration to `go-redis/redis` makes sense. So some additional understanding required here to proceed with changes.

I have not looked at code closely yet – will do at some point, for now some top-level things:

1. What was the reason of bump to `go1.18` in go.mod? We should not depend on features of `go1.18` for now.
2. The difference in timeout behavior. In the current implementation we have both read/write connection timeouts set, probably you can describe the difference in timeout behavior with new implementation. I see lot's of usages of `context.Background()` - does the operations have some underlying timeouts anyway? So some description about working with timeouts in implementation with `go-redis` would be awesome to have in PR.
3. Run existing benchmarks - we have several Redis benchmarks and they should perform better to justify the migration. Could you run them with https://pkg.go.dev/golang.org/x/perf/cmd/benchstat ? Super curious to look at results!
4. Performance in Redis Cluster – this is not covered by current benchmarks but we need to understand the performance difference in Cluster. Possibly we need some new benchmarks that measure how Engine performs in Cluster. Should not be worse than Redigo + `mna/redisc` either.
5. Have you tested failover with Sentinel? I think existing tests do not cover it - only connect to current master. Probably we can automate this somehow, but at least manual testing would be awesome to have some confidence it works in general. Otherwise, `redis-go` should be stable enough I believe to rely on it.
6. Though I think we still need to look through `go-redis` open issues list and try to find out whether some of bugs apply to Engine implementation.

Redis Cluster can be run with sth like this locally (but I suppose to compare the performance it's better to try excluding Docker from setup eventually):

```
version: "3.8"

services:
  cluster:
    image: redis:7.0.0-alpine
    entrypoint:
      - /bin/sh
      - -c
      - |
        redis-server --port 7001 --save "" --appendonly no --cluster-enabled yes --cluster-config-file 7001.conf &
        redis-server --port 7002 --save "" --appendonly no --cluster-enabled yes --cluster-config-file 7002.conf &
        redis-server --port 7003 --save "" --appendonly no --cluster-enabled yes --cluster-config-file 7003.conf &
        while ! redis-cli --cluster create 127.0.0.1:7001 127.0.0.1:7002 127.0.0.1:7003 --cluster-yes; do sleep 1; done
        wait
    ports:
      - "7001:7001"
      - "7002:7002"
      - "7003:7003"
```