@FZambia 2022-08-01
Hmm, interesting – I expected a better performance than with Redigo. Actually I did some measurements before opening that issue and results were quite promising. That was the biggest motivation to try the migration. So I'd say looking at them is the first step here do decide whether it worth or not. Since otherwise current implementation is pretty stable I believe to change it for sth new.