@j178 2022-08-02
Here are some benchmark results ran in my local machine:

goos: darwin
goarch: amd64
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz

Most numbers are fine, except for `RedisAddPresence_1Ch`, `RedisPresence_1Ch` and `RedisPresence_ManyCh`, especially `RedisAddPresence_1Ch` became incredibly slower. No idea why, still investigating.

```
name                     old time/op    new time/op    delta
RedisExtractPushData-12    54.9ns ± 0%    38.3ns ± 5%   -30.12%  (p=0.008 n=5+5)
name                     old alloc/op   new alloc/op   delta
RedisExtractPushData-12     16.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
name                     old allocs/op  new allocs/op  delta
RedisExtractPushData-12      1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)

// 100 nodes always stuck, it's wired, just skipped it for now
name                          old time/op  new time/op  delta
RedisSurvey/1_node_512B-12    1.54µs ±11%  1.30µs ± 6%  -15.77%  (p=0.008 n=5+5)
RedisSurvey/2_nodes_512B-12   49.7µs ± 2%  29.4µs ± 4%  -40.94%  (p=0.008 n=5+5)
RedisSurvey/3_nodes_512B-12   66.0µs ±15%  45.0µs ± 4%  -31.85%  (p=0.008 n=5+5)
RedisSurvey/4_nodes_512B-12   86.6µs ±11%  61.5µs ± 5%  -28.94%  (p=0.008 n=5+5)
RedisSurvey/5_nodes_512B-12    103µs ± 5%    82µs ± 3%  -20.66%  (p=0.008 n=5+5)
RedisSurvey/10_nodes_512B-12   246µs ±16%   200µs ±14%  -18.98%  (p=0.032 n=5+5)
RedisSurvey/1_node_4KB-12     1.86µs ± 8%  1.68µs ± 4%   -9.29%  (p=0.016 n=5+5)
RedisSurvey/2_nodes_4KB-12    77.3µs ±13%  57.4µs ±11%  -25.71%  (p=0.008 n=5+5)
RedisSurvey/3_nodes_4KB-12     180µs ±31%   173µs ±17%     ~     (p=1.000 n=5+5)
RedisSurvey/4_nodes_4KB-12     297µs ±55%   256µs ±29%     ~     (p=0.421 n=5+5)
RedisSurvey/5_nodes_4KB-12     393µs ±42%   328µs ±12%     ~     (p=0.151 n=5+5)
RedisSurvey/10_nodes_4KB-12    835µs ± 8%   473µs ±42%  -43.36%  (p=0.008 n=5+5)

name                     old time/op  new time/op  delta
RedisConsistentIndex-12  55.7ns ± 1%  51.5ns ± 0%  -7.53%  (p=0.016 n=5+4)

name           old time/op  new time/op  delta
RedisIndex-12  43.3ns ± 1%  39.5ns ± 2%  -8.70%  (p=0.008 n=5+5)

name                 old time/op  new time/op  delta
RedisPublish_1Ch-12  2.67µs ±12%  2.81µs ± 2%   ~     (p=0.222 n=5+5)

name                    old time/op  new time/op  delta
RedisPublish_ManyCh-12  2.42µs ± 1%  2.76µs ± 1%  +14.15%  (p=0.008 n=5+5)

name                                 old time/op  new time/op  delta
RedisPublish_History_1Ch/lists-12    27.7µs ± 2%  27.2µs ± 1%  -1.97%  (p=0.008 n=5+5)
RedisPublish_History_1Ch/streams-12  34.1µs ± 1%  33.8µs ± 1%    ~     (p=0.056 n=5+5)

name                                old time/op  new time/op  delta
RedisPub_History_ManyCh/lists-12    28.4µs ± 2%  28.0µs ± 5%   ~     (p=0.310 n=5+5)
RedisPub_History_ManyCh/streams-12  28.7µs ± 2%  28.6µs ± 5%   ~     (p=0.690 n=5+5)

name               old time/op  new time/op  delta
RedisSubscribe-12  2.20µs ± 7%  2.41µs ± 9%   ~     (p=0.056 n=5+5)

name                         old time/op  new time/op  delta
RedisHistory_1Ch/lists-12    23.0µs ± 8%  20.7µs ± 1%  -10.09%  (p=0.008 n=5+5)
RedisHistory_1Ch/streams-12  37.6µs ± 1%  36.8µs ± 0%   -2.17%  (p=0.008 n=5+5)

name                         old time/op  new time/op  delta
RedisRecover_1Ch/lists-12     388µs ± 3%   386µs ± 3%    ~     (p=0.421 n=5+5)
RedisRecover_1Ch/streams-12  48.8µs ± 2%  47.3µs ± 1%  -2.98%  (p=0.008 n=5+5)

name                              old time/op  new time/op  delta
RedisHistoryIteration/lists-12     591ms ± 2%   524ms ± 2%  -11.40%  (p=0.008 n=5+5)
RedisHistoryIteration/streams-12  57.1ms ± 2%  52.5ms ± 4%   -8.13%  (p=0.008 n=5+5)

// RedisAddPresence is awfully slower, still investigaing
name                     old time/op  new time/op    delta
RedisAddPresence_1Ch-12  15.8µs ± 0%  4680.2µs ± 1%  +29449.22%  (p=0.016 n=4+5)

name                  old time/op  new time/op  delta
RedisPresence_1Ch-12  13.8µs ± 1%  18.4µs ± 6%  +33.55%  (p=0.008 n=5+5)

name                     old time/op  new time/op  delta
RedisPresence_ManyCh-12  11.7µs ± 0%  15.7µs ± 4%  +33.82%  (p=0.008 n=5+5)
```