@j178 2022-08-02
After fixed the `redis.Nil` problem, here is the updated benchmark results about `RedisPresence`:
```
name                     old time/op  new time/op  delta
RedisAddPresence_1Ch-12  16.2µs ± 2%  16.6µs ±12%   ~     (p=0.690 n=5+5)

name                  old time/op  new time/op  delta
RedisPresence_1Ch-12  13.9µs ± 1%  12.9µs ± 1%  -7.44%  (p=0.008 n=5+5)

name                     old time/op  new time/op  delta
RedisPresence_ManyCh-12  11.9µs ± 1%  11.0µs ± 1%  -7.44%  (p=0.008 n=5+5)
```